Matter update hierarchy of plugins (#19915)
