+++ lib/libesp32/berry_matter/src/be_matter_module.c
@@ -210,7 +210,7 @@ extern const bclass be_class_Matter_TLV;   // need to declare it upfront because
 #include "../generate/be_matter_certs.h"
 
 #include "solidify/solidified_Matter_Plugin_1_Root.h"
-#include "solidify/solidified_Matter_Plugin_2_Aggregator.h"
+#include "solidify/solidified_Matter_Plugin_1_Aggregator.h"
 #include "solidify/solidified_Matter_Plugin_1_Device.h"
 #include "solidify/solidified_Matter_Plugin_2_OnOff.h"
 #include "solidify/solidified_Matter_Plugin_9_Virt_OnOff.h"
@@ -235,9 +235,9 @@ extern const bclass be_class_Matter_TLV;   // need to declare it upfront because
 #include "solidify/solidified_Matter_Plugin_9_Virt_Sensor_Illuminance.h"
 #include "solidify/solidified_Matter_Plugin_3_Sensor_Humidity.h"
 #include "solidify/solidified_Matter_Plugin_9_Virt_Sensor_Humidity.h"
-#include "solidify/solidified_Matter_Plugin_3_Sensor_Occupancy.h"
-#include "solidify/solidified_Matter_Plugin_3_Sensor_OnOff.h"
-#include "solidify/solidified_Matter_Plugin_3_Sensor_Contact.h"
+#include "solidify/solidified_Matter_Plugin_2_Sensor_Occupancy.h"
+#include "solidify/solidified_Matter_Plugin_2_Sensor_OnOff.h"
+#include "solidify/solidified_Matter_Plugin_2_Sensor_Contact.h"
 #include "solidify/solidified_Matter_Plugin_9_Virt_Sensor_Contact.h"
 #include "solidify/solidified_Matter_Plugin_9_Virt_Sensor_Occupancy.h"
 #include "solidify/solidified_Matter_Plugin_2_Bridge_HTTP.h"
@@ -251,8 +251,8 @@ extern const bclass be_class_Matter_TLV;   // need to declare it upfront because
 #include "solidify/solidified_Matter_Plugin_4_Bridge_Sensor_Temp.h"
 #include "solidify/solidified_Matter_Plugin_4_Bridge_Sensor_Illuminance.h"
 #include "solidify/solidified_Matter_Plugin_4_Bridge_Sensor_Humidity.h"
-#include "solidify/solidified_Matter_Plugin_4_Bridge_Sensor_Occupancy.h"
-#include "solidify/solidified_Matter_Plugin_4_Bridge_Sensor_Contact.h"
+#include "solidify/solidified_Matter_Plugin_3_Bridge_Sensor_Occupancy.h"
+#include "solidify/solidified_Matter_Plugin_3_Bridge_Sensor_Contact.h"
 #include "solidify/solidified_Matter_Plugin_4_Bridge_Sensor_Flow.h"
 
 /*********************************************************************************************\
