+++ lib/libesp32/berry_matter/src/solidify/solidified_Matter_Plugin_3_Bridge_Sensor_Occupancy.h
@@ -1,15 +1,15 @@
-/* Solidification of Matter_Plugin_4_Bridge_Sensor_Contact.h */
+/* Solidification of Matter_Plugin_3_Bridge_Sensor_Occupancy.h */
 /********************************************************************\
 * Generated code, don't edit                                         *
 \********************************************************************/
 #include "be_constobj.h"
 
-extern const bclass be_class_Matter_Plugin_Bridge_Sensor_Contact;
+extern const bclass be_class_Matter_Plugin_Bridge_Sensor_Occupancy;
 
 /********************************************************************
 ** Solidified function: <lambda>
 ********************************************************************/
-be_local_closure(Matter_Plugin_Bridge_Sensor_Contact__X3Clambda_X3E,   /* name */
+be_local_closure(Matter_Plugin_Bridge_Sensor_Occupancy__X3Clambda_X3E,   /* name */
   be_nested_proto(
     3,                          /* nstack */
     1,                          /* argc */
@@ -36,7 +36,7 @@ be_local_closure(Matter_Plugin_Bridge_Sensor_Contact__X3Clambda_X3E,   /* name *
 /********************************************************************
 ** Solidified function: init
 ********************************************************************/
-be_local_closure(Matter_Plugin_Bridge_Sensor_Contact_init,   /* name */
+be_local_closure(Matter_Plugin_Bridge_Sensor_Occupancy_init,   /* name */
   be_nested_proto(
     9,                          /* nstack */
     4,                          /* argc */
@@ -84,9 +84,68 @@ be_local_closure(Matter_Plugin_Bridge_Sensor_Contact_init,   /* name */
 
 
 /********************************************************************
+** Solidified function: parse_update
+********************************************************************/
+be_local_closure(Matter_Plugin_Bridge_Sensor_Occupancy_parse_update,   /* name */
+  be_nested_proto(
+    8,                          /* nstack */
+    3,                          /* argc */
+    2,                          /* varg */
+    0,                          /* has upvals */
+    NULL,                       /* no upvals */
+    0,                          /* has sup protos */
+    NULL,                       /* no sub protos */
+    1,                          /* has constants */
+    ( &(const bvalue[ 7]) {     /* constants */
+    /* K0   */  be_nested_str_weak(find),
+    /* K1   */  be_nested_str_weak(Switch),
+    /* K2   */  be_nested_str_weak(tasmota_switch_index),
+    /* K3   */  be_nested_str_weak(ON),
+    /* K4   */  be_nested_str_weak(shadow_occupancy),
+    /* K5   */  be_nested_str_weak(attribute_updated),
+    /* K6   */  be_const_int(0),
+    }),
+    be_str_weak(parse_update),
+    &be_const_str_solidified,
+    ( &(const binstruction[28]) {  /* code */
+      0x540E0007,  //  0000  LDINT	R3	8
+      0x1C0C0403,  //  0001  EQ	R3	R2	R3
+      0x780E0017,  //  0002  JMPF	R3	#001B
+      0x500C0000,  //  0003  LDBOOL	R3	0	0
+      0x8C100300,  //  0004  GETMET	R4	R1	K0
+      0x60180008,  //  0005  GETGBL	R6	G8
+      0x881C0102,  //  0006  GETMBR	R7	R0	K2
+      0x7C180200,  //  0007  CALL	R6	1
+      0x001A0206,  //  0008  ADD	R6	K1	R6
+      0x7C100400,  //  0009  CALL	R4	2
+      0x1C100903,  //  000A  EQ	R4	R4	K3
+      0x5C0C0800,  //  000B  MOVE	R3	R4
+      0x88100104,  //  000C  GETMBR	R4	R0	K4
+      0x4C140000,  //  000D  LDNIL	R5
+      0x20100805,  //  000E  NE	R4	R4	R5
+      0x78120009,  //  000F  JMPF	R4	#001A
+      0x88100104,  //  0010  GETMBR	R4	R0	K4
+      0x60140017,  //  0011  GETGBL	R5	G23
+      0x5C180600,  //  0012  MOVE	R6	R3
+      0x7C140200,  //  0013  CALL	R5	1
+      0x20100805,  //  0014  NE	R4	R4	R5
+      0x78120003,  //  0015  JMPF	R4	#001A
+      0x8C100105,  //  0016  GETMET	R4	R0	K5
+      0x541A0405,  //  0017  LDINT	R6	1030
+      0x581C0006,  //  0018  LDCONST	R7	K6
+      0x7C100600,  //  0019  CALL	R4	3
+      0x90020803,  //  001A  SETMBR	R0	K4	R3
+      0x80000000,  //  001B  RET	0
+    })
+  )
+);
+/*******************************************************************/
+
+
+/********************************************************************
 ** Solidified function: web_values_prefix
 ********************************************************************/
-be_local_closure(Matter_Plugin_Bridge_Sensor_Contact_web_values_prefix,   /* name */
+be_local_closure(Matter_Plugin_Bridge_Sensor_Occupancy_web_values_prefix,   /* name */
   be_nested_proto(
     10,                          /* nstack */
     1,                          /* argc */
@@ -138,68 +197,9 @@ be_local_closure(Matter_Plugin_Bridge_Sensor_Contact_web_values_prefix,   /* nam
 
 
 /********************************************************************
-** Solidified function: parse_update
-********************************************************************/
-be_local_closure(Matter_Plugin_Bridge_Sensor_Contact_parse_update,   /* name */
-  be_nested_proto(
-    8,                          /* nstack */
-    3,                          /* argc */
-    2,                          /* varg */
-    0,                          /* has upvals */
-    NULL,                       /* no upvals */
-    0,                          /* has sup protos */
-    NULL,                       /* no sub protos */
-    1,                          /* has constants */
-    ( &(const bvalue[ 7]) {     /* constants */
-    /* K0   */  be_nested_str_weak(find),
-    /* K1   */  be_nested_str_weak(Switch),
-    /* K2   */  be_nested_str_weak(tasmota_switch_index),
-    /* K3   */  be_nested_str_weak(ON),
-    /* K4   */  be_nested_str_weak(shadow_contact),
-    /* K5   */  be_nested_str_weak(attribute_updated),
-    /* K6   */  be_const_int(0),
-    }),
-    be_str_weak(parse_update),
-    &be_const_str_solidified,
-    ( &(const binstruction[28]) {  /* code */
-      0x540E0007,  //  0000  LDINT	R3	8
-      0x1C0C0403,  //  0001  EQ	R3	R2	R3
-      0x780E0017,  //  0002  JMPF	R3	#001B
-      0x500C0000,  //  0003  LDBOOL	R3	0	0
-      0x8C100300,  //  0004  GETMET	R4	R1	K0
-      0x60180008,  //  0005  GETGBL	R6	G8
-      0x881C0102,  //  0006  GETMBR	R7	R0	K2
-      0x7C180200,  //  0007  CALL	R6	1
-      0x001A0206,  //  0008  ADD	R6	K1	R6
-      0x7C100400,  //  0009  CALL	R4	2
-      0x1C100903,  //  000A  EQ	R4	R4	K3
-      0x5C0C0800,  //  000B  MOVE	R3	R4
-      0x88100104,  //  000C  GETMBR	R4	R0	K4
-      0x4C140000,  //  000D  LDNIL	R5
-      0x20100805,  //  000E  NE	R4	R4	R5
-      0x78120009,  //  000F  JMPF	R4	#001A
-      0x88100104,  //  0010  GETMBR	R4	R0	K4
-      0x60140017,  //  0011  GETGBL	R5	G23
-      0x5C180600,  //  0012  MOVE	R6	R3
-      0x7C140200,  //  0013  CALL	R5	1
-      0x20100805,  //  0014  NE	R4	R4	R5
-      0x78120003,  //  0015  JMPF	R4	#001A
-      0x8C100105,  //  0016  GETMET	R4	R0	K5
-      0x541A0044,  //  0017  LDINT	R6	69
-      0x581C0006,  //  0018  LDCONST	R7	K6
-      0x7C100600,  //  0019  CALL	R4	3
-      0x90020803,  //  001A  SETMBR	R0	K4	R3
-      0x80000000,  //  001B  RET	0
-    })
-  )
-);
-/*******************************************************************/
-
-
-/********************************************************************
 ** Solidified function: read_attribute
 ********************************************************************/
-be_local_closure(Matter_Plugin_Bridge_Sensor_Contact_read_attribute,   /* name */
+be_local_closure(Matter_Plugin_Bridge_Sensor_Occupancy_read_attribute,   /* name */
   be_nested_proto(
     12,                          /* nstack */
     4,                          /* argc */
@@ -209,30 +209,32 @@ be_local_closure(Matter_Plugin_Bridge_Sensor_Contact_read_attribute,   /* name *
     0,                          /* has sup protos */
     NULL,                       /* no sub protos */
     1,                          /* has constants */
-    ( &(const bvalue[12]) {     /* constants */
+    ( &(const bvalue[14]) {     /* constants */
     /* K0   */  be_nested_str_weak(matter),
     /* K1   */  be_nested_str_weak(TLV),
     /* K2   */  be_nested_str_weak(cluster),
     /* K3   */  be_nested_str_weak(attribute),
     /* K4   */  be_const_int(0),
-    /* K5   */  be_nested_str_weak(shadow_contact),
+    /* K5   */  be_nested_str_weak(shadow_occupancy),
     /* K6   */  be_nested_str_weak(set),
-    /* K7   */  be_nested_str_weak(BOOL),
+    /* K7   */  be_nested_str_weak(U1),
     /* K8   */  be_nested_str_weak(NULL),
-    /* K9   */  be_nested_str_weak(U4),
-    /* K10  */  be_const_int(1),
-    /* K11  */  be_nested_str_weak(read_attribute),
+    /* K9   */  be_const_int(1),
+    /* K10  */  be_const_int(3),
+    /* K11  */  be_const_int(2),
+    /* K12  */  be_nested_str_weak(U4),
+    /* K13  */  be_nested_str_weak(read_attribute),
     }),
     be_str_weak(read_attribute),
     &be_const_str_solidified,
-    ( &(const binstruction[53]) {  /* code */
+    ( &(const binstruction[69]) {  /* code */
       0xB8120000,  //  0000  GETNGBL	R4	K0
       0x88100901,  //  0001  GETMBR	R4	R4	K1
       0x88140502,  //  0002  GETMBR	R5	R2	K2
       0x88180503,  //  0003  GETMBR	R6	R2	K3
-      0x541E0044,  //  0004  LDINT	R7	69
+      0x541E0405,  //  0004  LDINT	R7	1030
       0x1C1C0A07,  //  0005  EQ	R7	R5	R7
-      0x781E0023,  //  0006  JMPF	R7	#002B
+      0x781E0033,  //  0006  JMPF	R7	#003B
       0x1C1C0D04,  //  0007  EQ	R7	R6	K4
       0x781E000F,  //  0008  JMPF	R7	#0019
       0x881C0105,  //  0009  GETMBR	R7	R0	K5
@@ -250,35 +252,51 @@ be_local_closure(Matter_Plugin_Bridge_Sensor_Contact_read_attribute,   /* name *
       0x4C280000,  //  0015  LDNIL	R10
       0x7C1C0600,  //  0016  CALL	R7	3
       0x80040E00,  //  0017  RET	1	R7
-      0x70020010,  //  0018  JMP		#002A
-      0x541EFFFB,  //  0019  LDINT	R7	65532
-      0x1C1C0C07,  //  001A  EQ	R7	R6	R7
-      0x781E0005,  //  001B  JMPF	R7	#0022
-      0x8C1C0706,  //  001C  GETMET	R7	R3	K6
-      0x88240909,  //  001D  GETMBR	R9	R4	K9
-      0x58280004,  //  001E  LDCONST	R10	K4
-      0x7C1C0600,  //  001F  CALL	R7	3
-      0x80040E00,  //  0020  RET	1	R7
-      0x70020007,  //  0021  JMP		#002A
-      0x541EFFFC,  //  0022  LDINT	R7	65533
-      0x1C1C0C07,  //  0023  EQ	R7	R6	R7
-      0x781E0004,  //  0024  JMPF	R7	#002A
-      0x8C1C0706,  //  0025  GETMET	R7	R3	K6
-      0x88240909,  //  0026  GETMBR	R9	R4	K9
-      0x5828000A,  //  0027  LDCONST	R10	K10
-      0x7C1C0600,  //  0028  CALL	R7	3
-      0x80040E00,  //  0029  RET	1	R7
-      0x70020008,  //  002A  JMP		#0034
-      0x601C0003,  //  002B  GETGBL	R7	G3
-      0x5C200000,  //  002C  MOVE	R8	R0
-      0x7C1C0200,  //  002D  CALL	R7	1
-      0x8C1C0F0B,  //  002E  GETMET	R7	R7	K11
-      0x5C240200,  //  002F  MOVE	R9	R1
-      0x5C280400,  //  0030  MOVE	R10	R2
-      0x5C2C0600,  //  0031  MOVE	R11	R3
-      0x7C1C0800,  //  0032  CALL	R7	4
-      0x80040E00,  //  0033  RET	1	R7
-      0x80000000,  //  0034  RET	0
+      0x70020020,  //  0018  JMP		#003A
+      0x1C1C0D09,  //  0019  EQ	R7	R6	K9
+      0x781E0005,  //  001A  JMPF	R7	#0021
+      0x8C1C0706,  //  001B  GETMET	R7	R3	K6
+      0x88240907,  //  001C  GETMBR	R9	R4	K7
+      0x5828000A,  //  001D  LDCONST	R10	K10
+      0x7C1C0600,  //  001E  CALL	R7	3
+      0x80040E00,  //  001F  RET	1	R7
+      0x70020018,  //  0020  JMP		#003A
+      0x1C1C0D0B,  //  0021  EQ	R7	R6	K11
+      0x781E0005,  //  0022  JMPF	R7	#0029
+      0x8C1C0706,  //  0023  GETMET	R7	R3	K6
+      0x88240907,  //  0024  GETMBR	R9	R4	K7
+      0x58280004,  //  0025  LDCONST	R10	K4
+      0x7C1C0600,  //  0026  CALL	R7	3
+      0x80040E00,  //  0027  RET	1	R7
+      0x70020010,  //  0028  JMP		#003A
+      0x541EFFFB,  //  0029  LDINT	R7	65532
+      0x1C1C0C07,  //  002A  EQ	R7	R6	R7
+      0x781E0005,  //  002B  JMPF	R7	#0032
+      0x8C1C0706,  //  002C  GETMET	R7	R3	K6
+      0x8824090C,  //  002D  GETMBR	R9	R4	K12
+      0x58280004,  //  002E  LDCONST	R10	K4
+      0x7C1C0600,  //  002F  CALL	R7	3
+      0x80040E00,  //  0030  RET	1	R7
+      0x70020007,  //  0031  JMP		#003A
+      0x541EFFFC,  //  0032  LDINT	R7	65533
+      0x1C1C0C07,  //  0033  EQ	R7	R6	R7
+      0x781E0004,  //  0034  JMPF	R7	#003A
+      0x8C1C0706,  //  0035  GETMET	R7	R3	K6
+      0x8824090C,  //  0036  GETMBR	R9	R4	K12
+      0x5828000A,  //  0037  LDCONST	R10	K10
+      0x7C1C0600,  //  0038  CALL	R7	3
+      0x80040E00,  //  0039  RET	1	R7
+      0x70020008,  //  003A  JMP		#0044
+      0x601C0003,  //  003B  GETGBL	R7	G3
+      0x5C200000,  //  003C  MOVE	R8	R0
+      0x7C1C0200,  //  003D  CALL	R7	1
+      0x8C1C0F0D,  //  003E  GETMET	R7	R7	K13
+      0x5C240200,  //  003F  MOVE	R9	R1
+      0x5C280400,  //  0040  MOVE	R10	R2
+      0x5C2C0600,  //  0041  MOVE	R11	R3
+      0x7C1C0800,  //  0042  CALL	R7	4
+      0x80040E00,  //  0043  RET	1	R7
+      0x80000000,  //  0044  RET	0
     })
   )
 );
@@ -288,7 +306,7 @@ be_local_closure(Matter_Plugin_Bridge_Sensor_Contact_read_attribute,   /* name *
 /********************************************************************
 ** Solidified function: web_values
 ********************************************************************/
-be_local_closure(Matter_Plugin_Bridge_Sensor_Contact_web_values,   /* name */
+be_local_closure(Matter_Plugin_Bridge_Sensor_Occupancy_web_values,   /* name */
   be_nested_proto(
     10,                          /* nstack */
     1,                          /* argc */
@@ -302,10 +320,10 @@ be_local_closure(Matter_Plugin_Bridge_Sensor_Contact_web_values,   /* name */
     /* K0   */  be_nested_str_weak(webserver),
     /* K1   */  be_nested_str_weak(web_values_prefix),
     /* K2   */  be_nested_str_weak(content_send),
-    /* K3   */  be_nested_str_weak(Contact_X25i_X20_X25s),
+    /* K3   */  be_nested_str_weak(Occupancy_X25i_X20_X25s),
     /* K4   */  be_nested_str_weak(tasmota_switch_index),
     /* K5   */  be_nested_str_weak(web_value_onoff),
-    /* K6   */  be_nested_str_weak(shadow_contact),
+    /* K6   */  be_nested_str_weak(shadow_occupancy),
     }),
     be_str_weak(web_values),
     &be_const_str_solidified,
@@ -330,25 +348,28 @@ be_local_closure(Matter_Plugin_Bridge_Sensor_Contact_web_values,   /* name */
 
 
 /********************************************************************
-** Solidified class: Matter_Plugin_Bridge_Sensor_Contact
+** Solidified class: Matter_Plugin_Bridge_Sensor_Occupancy
 ********************************************************************/
 extern const bclass be_class_Matter_Plugin_Bridge_HTTP;
-be_local_class(Matter_Plugin_Bridge_Sensor_Contact,
+be_local_class(Matter_Plugin_Bridge_Sensor_Occupancy,
     2,
     &be_class_Matter_Plugin_Bridge_HTTP,
     be_nested_map(16,
     ( (struct bmapnode*) &(const bmapnode[]) {
-        { be_const_key_weak(ARG_TYPE, 5), be_const_static_closure(Matter_Plugin_Bridge_Sensor_Contact__X3Clambda_X3E_closure) },
-        { be_const_key_weak(ARG_HINT, 4), be_nested_str_weak(Switch_X3Cx_X3E_X20number) },
-        { be_const_key_weak(shadow_contact, -1), be_const_var(1) },
-        { be_const_key_weak(init, -1), be_const_closure(Matter_Plugin_Bridge_Sensor_Contact_init_closure) },
-        { be_const_key_weak(web_values_prefix, -1), be_const_closure(Matter_Plugin_Bridge_Sensor_Contact_web_values_prefix_closure) },
-        { be_const_key_weak(web_values, -1), be_const_closure(Matter_Plugin_Bridge_Sensor_Contact_web_values_closure) },
-        { be_const_key_weak(DISPLAY_NAME, -1), be_nested_str_weak(Contact) },
+        { be_const_key_weak(ARG_TYPE, 4), be_const_static_closure(Matter_Plugin_Bridge_Sensor_Occupancy__X3Clambda_X3E_closure) },
+        { be_const_key_weak(ARG_HINT, 12), be_nested_str_weak(Switch_X3Cx_X3E_X20number) },
+        { be_const_key_weak(shadow_occupancy, -1), be_const_var(1) },
+        { be_const_key_weak(init, -1), be_const_closure(Matter_Plugin_Bridge_Sensor_Occupancy_init_closure) },
+        { be_const_key_weak(web_values, -1), be_const_closure(Matter_Plugin_Bridge_Sensor_Occupancy_web_values_closure) },
+        { be_const_key_weak(DISPLAY_NAME, -1), be_nested_str_weak(Occupancy) },
+        { be_const_key_weak(TYPES, 9), be_const_simple_instance(be_nested_simple_instance(&be_class_map, {
+        be_const_map( *     be_nested_map(1,
+    ( (struct bmapnode*) &(const bmapnode[]) {
+        { be_const_key_int(263, -1), be_const_int(2) },
+    }))    ) } )) },
         { be_const_key_weak(ARG, -1), be_nested_str_weak(switch) },
-        { be_const_key_weak(parse_update, 11), be_const_closure(Matter_Plugin_Bridge_Sensor_Contact_parse_update_closure) },
-        { be_const_key_weak(read_attribute, 12), be_const_closure(Matter_Plugin_Bridge_Sensor_Contact_read_attribute_closure) },
-        { be_const_key_weak(CLUSTERS, 9), be_const_simple_instance(be_nested_simple_instance(&be_class_map, {
+        { be_const_key_weak(parse_update, 14), be_const_closure(Matter_Plugin_Bridge_Sensor_Occupancy_parse_update_closure) },
+        { be_const_key_weak(CLUSTERS, 11), be_const_simple_instance(be_nested_simple_instance(&be_class_map, {
         be_const_map( *     be_nested_map(6,
     ( (struct bmapnode*) &(const bmapnode[]) {
         { be_const_key_int(5, -1), be_const_simple_instance(be_nested_simple_instance(&be_class_list, {
@@ -363,15 +384,14 @@ be_local_class(Matter_Plugin_Bridge_Sensor_Contact,
         be_const_int(65532),
         be_const_int(65533),
     }))    ) } )) },
-        { be_const_key_int(3, -1), be_const_simple_instance(be_nested_simple_instance(&be_class_list, {
-        be_const_list( *     be_nested_list(4,
+        { be_const_key_int(4, -1), be_const_simple_instance(be_nested_simple_instance(&be_class_list, {
+        be_const_list( *     be_nested_list(3,
     ( (struct bvalue*) &(const bvalue[]) {
         be_const_int(0),
-        be_const_int(1),
         be_const_int(65532),
         be_const_int(65533),
     }))    ) } )) },
-        { be_const_key_int(57, 1), be_const_simple_instance(be_nested_simple_instance(&be_class_list, {
+        { be_const_key_int(57, -1), be_const_simple_instance(be_nested_simple_instance(&be_class_list, {
         be_const_list( *     be_nested_list(7,
     ( (struct bvalue*) &(const bvalue[]) {
         be_const_int(17),
@@ -382,17 +402,20 @@ be_local_class(Matter_Plugin_Bridge_Sensor_Contact,
         be_const_int(17),
         be_const_int(18),
     }))    ) } )) },
-        { be_const_key_int(69, 2), be_const_simple_instance(be_nested_simple_instance(&be_class_list, {
-        be_const_list( *     be_nested_list(3,
+        { be_const_key_int(3, 2), be_const_simple_instance(be_nested_simple_instance(&be_class_list, {
+        be_const_list( *     be_nested_list(4,
     ( (struct bvalue*) &(const bvalue[]) {
         be_const_int(0),
+        be_const_int(1),
         be_const_int(65532),
         be_const_int(65533),
     }))    ) } )) },
-        { be_const_key_int(4, -1), be_const_simple_instance(be_nested_simple_instance(&be_class_list, {
-        be_const_list( *     be_nested_list(3,
+        { be_const_key_int(1030, 1), be_const_simple_instance(be_nested_simple_instance(&be_class_list, {
+        be_const_list( *     be_nested_list(5,
     ( (struct bvalue*) &(const bvalue[]) {
         be_const_int(0),
+        be_const_int(1),
+        be_const_int(2),
         be_const_int(65532),
         be_const_int(65533),
     }))    ) } )) },
@@ -407,23 +430,20 @@ be_local_class(Matter_Plugin_Bridge_Sensor_Contact,
         be_const_int(65533),
     }))    ) } )) },
     }))    ) } )) },
+        { be_const_key_weak(tasmota_switch_index, 6), be_const_var(0) },
+        { be_const_key_weak(read_attribute, -1), be_const_closure(Matter_Plugin_Bridge_Sensor_Occupancy_read_attribute_closure) },
+        { be_const_key_weak(web_values_prefix, -1), be_const_closure(Matter_Plugin_Bridge_Sensor_Occupancy_web_values_prefix_closure) },
+        { be_const_key_weak(TYPE, 5), be_nested_str_weak(http_occupancy) },
         { be_const_key_weak(UPDATE_TIME, -1), be_const_int(5000) },
-        { be_const_key_weak(TYPES, 14), be_const_simple_instance(be_nested_simple_instance(&be_class_map, {
-        be_const_map( *     be_nested_map(1,
-    ( (struct bmapnode*) &(const bmapnode[]) {
-        { be_const_key_int(21, -1), be_const_int(1) },
-    }))    ) } )) },
-        { be_const_key_weak(TYPE, 6), be_nested_str_weak(http_contact) },
-        { be_const_key_weak(tasmota_switch_index, -1), be_const_var(0) },
-        { be_const_key_weak(UPDATE_CMD, -1), be_nested_str_weak(Status_X208) },
+        { be_const_key_weak(UPDATE_CMD, 2), be_nested_str_weak(Status_X208) },
     })),
-    be_str_weak(Matter_Plugin_Bridge_Sensor_Contact)
+    be_str_weak(Matter_Plugin_Bridge_Sensor_Occupancy)
 );
 /*******************************************************************/
 
-void be_load_Matter_Plugin_Bridge_Sensor_Contact_class(bvm *vm) {
-    be_pushntvclass(vm, &be_class_Matter_Plugin_Bridge_Sensor_Contact);
-    be_setglobal(vm, "Matter_Plugin_Bridge_Sensor_Contact");
+void be_load_Matter_Plugin_Bridge_Sensor_Occupancy_class(bvm *vm) {
+    be_pushntvclass(vm, &be_class_Matter_Plugin_Bridge_Sensor_Occupancy);
+    be_setglobal(vm, "Matter_Plugin_Bridge_Sensor_Occupancy");
     be_pop(vm, 1);
 }
 /********************************************************************/
